backend/requirements/base.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-extensions==3.2.3
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pillow==10.1.0
python-decouple==3.8
django-redis==5.4.0
drf-spectacular==0.26.5
python-dateutil==2.8.2
requests==2.31.0
stripe==7.0.0
django-storages==1.14.2
boto3==1.34.0
